feat: create generate method for token

+++ services/auth-service/internal/domain/repository_interface/token_repository_interface.go
@@ -9,7 +9,7 @@ type TokenRepositoryInterface interface {
 	Create(ctx context.Context, token *entity.Token) error
 	FindByToken(ctx context.Context, token string) (*entity.Token, error)
 	Revoke(ctx context.Context, token string) error
-	FindByUserId(ctx context.Context, userId int64) (*entity.Token, error)
+	FindByUserId(ctx context.Context, userId int) (*entity.Token, error)
 	Delete(ctx context.Context, id int) error
-	RevokeAllUserTokens(ctx context.Context, userId int64) error
+	RevokeAllUserTokens(ctx context.Context, userId int) error
 }
